levels: Fix GRAY1 walls around the trilobite
In the somewhat triangular room there is a trilobite below the
surrounding floor in sector 84. The trilobite seems to pop into
existence as the player approaches. As a result of the Eureka cleanup
the trilobite now appears to be surrounded by GRAY1 walls. The solution
is to make the room that contains the trilobite as small as possible
(length 2 per side) and to use a linedef texture that is close to the
floor (WOOD5). It would take an extremely observant player to notice the
tiny wood colored hole in the wood floor.