levels: Fix GRAY1 walls around the end trilobites
This is similar to the previous fix. Prior to this fix the player
would have to be pressed up against the closest corner of the balcony
across from the exit to notice. Now it's all but impossible to notice.

The trilobites are raised before the player enters the room with the
exit portal, so it doesn't matter how visible the hole is in that room.

Also, there is a similar issue with a hole that contains 4 hatchlings.
In that case since the switch that raises them is far away there's
not much reason to fix it. I just used a less noticable texture. To an
observant player it just looks like a hole in the floor that they could
pass over.