levels: DM08 Restore the circular saw blades
Prior to the Eureka cleanup DM08 had circular saw blades on the left
side of the map via texture SLAD2. Unfortunately Eureka falsely flags
SLAD2 as a Medusa causing textures for middle textures since it
contains more than one patch. This fixes introduces texture SAW2 which
only contains the SAW2 patch, and which therefore does not cause the
Eureka warning. It should look exactly the same to the player.