dehacked: fix some map names, revise one more.
DM21 typo; DM12 use the most neutral version of that abbreviation.

Map11: it is neither infinite nor a plain. Inasmuch as it's a misleading name it's arguably a great name for a map that's basically a ceaseless parade of misleading gotchas (fake doors, switches requiring multiple backtracks to find out what you need to do to progress, secret areas that kill you, etc.) but it just feels like this name used to be for a different map (maybe a followup to the immediately preceding Outlands) and no one bothered to update it.

Someone who actually enjoys this map please feel free to suggest a different name.