sprites: [BUGFIX] make torch frames the same size.
Currently the sprites are all cropped to exactly as much as needed. Apparently this has some really bad side effects with the animations.

To check: warp to Map29 in vanilla and take a couple steps back. Expected: solid parts of torches stay perfectly still except for the shifting lights on them. Actual: torches jitter and twitch.

Making every sprite in each loop exactly the same size gets rid of this problem.