sprites: better match shotgun pickups to hud.
The barrel shroud on SHOTA0 has been off model for a long time - I didn't really want to touch it back in 2019 because I wasn't really sure what was even going on in the sprite, but now that I've been looking at the shotgun a lot more it's pretty clear it could be improved.

Also made the shells in the stock wider and looking a bit more like what we see in the shell pickup. (This entire thing had actually started off as a simple attempt to change the colour of the shells to the olive green the shell pickup uses.)

The SSG pickup has also felt "off" to me for a long time, but couldn't really put my finger on it other than "the foregrip's too big and i don't like it". Turns out if it "really" were that big you should see it in at least one of the reloading frames.

The raised rear... bead? peg? was also implemented and the vents on the end of the barrel made a bit further apart to factor in the foreshortening of the HUD sprite.