levels: overhaul E3M1 start.
First, the player had started off looking away from the serpentipede that could see them right away, which felt like I was being railroaded into a "David Madsen in the Dark Room" situation. It turns out that I can't quite move the start far back enough to get all players out of LOS without changing the map geometry, which leads to...

Second, the player starts on a teleporter pad *and* there's an implied trail behind them - which way had the player come? So to address that I've gotten rid of the tiny little pad and instead created this huge elaborate teleporter setup that could reasonably look like the destination from the interplanetary teleporter described at the end of Episode 2.

Also stashed a dead zombieman in the back, so the implied immediate backstory is that Karl warped into Horizon, saw a zombieman, killed it before it could attack or sound an alarm, then dragged the body into this crevice and tried to stay out of sight, before realizing he couldn't climb back out that way...