sprites: upgrade health pickups.
Korp's medikit and stimpack as shared on the Discord, recoloured to use the Doom palette.

The health bonus originally started out as my recolour of Korp's green bottle to give it the same yellow colour coding as the hyperdrive sphere. (If it looks too much like the medical pickups the health going over 100% looks like a bug.) The design I ended up with is inspired by Pac-Man's dots and Unreal's health plants.

The berserk is reshaded and given a similar cylindrical look to the stimpack. The face is no longer a human skull but something more alien.