Edit to E1M2 (phase 1) (#851)
* Update e1m2.wad

* Update e1m2.wad

Changed two encounters with Pinkies to not be closed arenas (the one when you get the blue key, and the one right before the end of the map)

* Update e1m2.wad

Visual edits by mc776. Also I added a new secret right before the end and made some visual tweaks to better indicate other secrets.

* Update e1m2.wad

Edits by mc776 and minor by me, changing the final outdoor area to be bigger and more interesting, changed the lift to the secret in sector 64, and fixed misalignment of step texture of linedef 428.