patches: normalize crate shading. (#884)
Issue #866 turns out to be the case for all of these.

The little wood crates and VGCRATE1 are already normal in their shading.

The big wood crates are a mix of contiguous or outlined on both sides; the wood grain means any such stuff isn't very visible anyways.

I've also updated *CRATEM1 to give them the dark lined look the ends of the crates have. This might make the repetition too obvious but overall it seems like an improvement.