Add files via upload
**Map01**
Swapped out the impassable line for narrowing the big door gap enough for it to actually clearly visibly block the player's movement.

Used the flat bleeding trick to get rid of that little spotlight under the exit sign. The distracting bottom light on the sign itself has also been removed.


**Map03**
Unflagged some invisible impassable lines in front of screens and switches in the first corridor since they weren't needed in the first place.

Changed floor/ceiling textures on the curvy screen corner so it wasn't blatantly conflicting with the tile pattern.

Added a railing onto the big nukage rooms.


**Map12**
Flyers in the far northwest room could not figure out how to get into the big arena if they're in the dead ends. The side doors can now be opened by them.

Fixed some misunpegged lowering platforms/item closets in the south.


**Map15**
Second attempt to fix the rising shipping elevator. No more moving bridge (at all - the new safety lifts are still there), both switches available again, everything actually takes some time and moves visibly now.

Tag 56 changed to the four lines on the border of the pit.

Moved the cells even deeper into the expanded secret room as parts of the hitbox were still sticking out on the other side of something.


**Map16**
Unaligned the MIDSPACE in the new back alley.


**Map21**
Added an implied ladder leading down into the starting room.

Fixed the plasma secret in the library. If you went inside the room and let it close behind you, there was no way out; the way you came is now given the same lift tag. The secret itself is marked with something better than just suspiciously secret-tagged linedefs on the automap. The library south wall is also no longer incorrectly tagged as not to show up on the map.

Fixed the misaligned, improperly pegged door side texture in the big wall that opens up to the north of the big lava pit.

The lift leading into that room overlooking the big lava pit can now be traversed both ways.

You can actually fall into the lava in (-948,1836) with no way out. Narrowed that section accordingly.


**Map23**
Reduced number of sectors and number of secret sectors (doorway into the actual secret also flagged as secret) in the southeastern secret.

Fixed lone floating rocket by the crate north of the G.


**Map25**
Northwest baron/chaingunner teleport ambush was double-triggered and one of them did the opposite of what it was supposed to do. That entire area has been redone as a creepy wooden shack, to get rid of the arbitrary anti-jump invisible blocking lines.

Also added a sector to make sure noises pass through to those two individual monsters.

Unstuck serpentipede # 444.

The exit room seemed a bit dark for that lava so I took the liberty of changing it to 5% and a hot (but solid) rock texture.

Overhauled the starting entrance to explain why you were stuck in a little cramped chamber that you could only exit by shooting something behind you.


**Map32**
Got rid of the double-sector red key lift, in favour of just the outside portion being the lift.

Fixed all the windows so you can't jump through them in GZDoom under normal circumstances. Added some decorative railings.

Marked the berserk secret so it can be found without checking the automap (and can no longer be found by checking the automap), but in that specifically annoying, obtuse way that everything else is intentionally done on this map.

(NOTE: these Map32 fixes are pending another major improvement to this map that rednakhla is working on and may be overridden by other decisions in their version. I'm just throwing this out there as a temporary fix as we do not appear to have a timeline for that version yet. If it is completed before this PR is merged I'm more than happy to add it here.)