levels: map21 fixes.
The lava groove border could be jumped into as it was exactly 32 units wide. Reduced this all around to 24, and added a damagefloor there just in case.

One lava damage floor had 20% for no apparent reason where everywhere else the player would be expecting it to do 10. It now does 10.

Rearranged a bunch of stuff in the final corridor for the bricks to line up.