flats: remove vert. lines in fbodies and flat5_6b.
There's no way they're going to tile with the equivalent wall textures and I don't think that's even a realistic goal, but at least it should be close enough not to be immediately noticeable in software mode (cf. when someone tries to use FLAT5_6 and the bloody skull wall, which is what prompted this to begin with).