levels: improve Map27.
It turns out DOGB01 had posted an improved version of the map on Discord back in April 2022 and it slipped past everyone. I've taken it and re-implemented the DM starts (and additional door lines to let people into the central room from the other direction) and fixed the lizard baby and candle rooms as before.