levels: merge Map27 sectors.
The symmetrical aesthetic of this map lends itself to a *lot* of non-contiguous areas with the exact same floor and ceiling heights, flats and lights. I've gone and merged a whole bunch of these wherever I select a cluster and don't see "<multiple values>" in any field in SLADE.

I've also just plain deleted the corner pillar sectors in the rectangular pinkie closet in the far west because they just seemed to be regular pillars.

Also lowered the starting room exit big skull door textures a bit, as from the starting side the top skull was clipping into the wall and on the other side it was a little hard to see the bottom of the skull (and distinguish it visually from the wood panelling nearby).

Posting this as a separate commit just in case I broke anything important.