levels: misc. Map27 fixes.
Room immediately north of the big cross room has these little cage things in the corners. The bar textures were not impassable and are made impassable now.

One of the columns in the final arena had inwards-facing linedefs causing rockets to disappear.

Deleted the sector representing the middle piece of the central pillar just south of the exit, as it's effectively indistinguishable from a set of one-sided linedefs (other than that it was making rockets disappear).

More sector minimizations.

Added one more torch to mark the lizardbaby secret, because frankly no one's going to think anything about a visibly misaligned texture on one of the easiest textures to accidentally visibly misalign.