levels: improve Map27. (#892)
* levels: improve Map27.

It turns out DOGB01 had posted an improved version of the map on Discord back in April 2022 and it slipped past everyone. I've taken it and re-implemented the DM starts (and additional door lines to let people into the central room from the other direction) and fixed the lizard baby and candle rooms as before.

* levels: merge Map27 sectors.

The symmetrical aesthetic of this map lends itself to a *lot* of non-contiguous areas with the exact same floor and ceiling heights, flats and lights. I've gone and merged a whole bunch of these wherever I select a cluster and don't see "<multiple values>" in any field in SLADE.

I've also just plain deleted the corner pillar sectors in the rectangular pinkie closet in the far west because they just seemed to be regular pillars.

Also lowered the starting room exit big skull door textures a bit, as from the starting side the top skull was clipping into the wall and on the other side it was a little hard to see the bottom of the skull (and distinguish it visually from the wood panelling nearby).

Posting this as a separate commit just in case I broke anything important.

* levels: misc. Map27 fixes.

Room immediately north of the big cross room has these little cage things in the corners. The bar textures were not impassable and are made impassable now.

One of the columns in the final arena had inwards-facing linedefs causing rockets to disappear.

Deleted the sector representing the middle piece of the central pillar just south of the exit, as it's effectively indistinguishable from a set of one-sided linedefs (other than that it was making rockets disappear).

More sector minimizations.

Added one more torch to mark the lizardbaby secret, because frankly no one's going to think anything about a visibly misaligned texture on one of the easiest textures to accidentally visibly misalign.

* map27: make caver navigable.

The Palace of Red is a meticulously detailed map in which I always end up spending the most time staring at minute details in the ugliest and dimmest part of it, just because I keep getting lost because there aren't any landmarks and I can't see past a few feet.

Here's a bit of redecorating:
- level changes exaggerated between the various prongs
- sediment buildup added to edges of blood river
- two different kinds of rock, the main one being much darker to give the impression of darkness without making the area unnavigable
- increased lighting depending on proximity to existing torches and skylight
- candles added to yellow door