levels: E1M6 aesthetic touchups.
After seeing that screenshot with the building with that side entirely in featureless grey I had a Reaction and started this.

I've also made some impassable lines visible for the benefit of jumpers. Basically anything where the bottom of the window is less than 64 units from the floor is jumpable and needs to be indicated as impassable. The training centre theme suggested liberal use of thrown-heavy-object-proof chain link fence.

I'm sorry but there is no rational explanation for those featureless polished silver walls in the outside-adjacent parts of that giant nukage chamber/pool. They're algae-ridden cement now.

Rest of the map is great. Love how it gives the actual vibe of a huge athletic training centre without going into awkwardly specific detail. Besides the shaped columns on that big window wall (and maybe visibility of attackers in the courtyard) there should be no changes that can potentially affect gameplay.