levels: map02 tweaks.
The yellow door appears to open by splitting itself into 3 parts, then moving only the middle part leaving the sides behind. This is now fixed so that the entire door moves.

Touched up the texture alignment above the exit door and replaced the floor on top of the monitor thingy blocking direct fire from the serpentipedes in front of the yellow door.

The little drops into the water area near the start can be cheesed by staying in the lit areas behind the monster blocking lines. This seems to be unintended behaviour so I've added new blocking lines that are exactly the size of the player.

The big drop past the red and blue doors needs something to stop the pinkies from blocking the player's initial drop. They're all set to +ambush now.

The exit sign has been given the flat-bleed treatment as Map01 and Map26, but this time I've actually kept the lower floor within monster step height. (I'd fix it for Map01 but it turns out the monsters step around it without issue.)