dehacked: tiny intertext tweaks.
The E4 ending text didn't need to split the phrase line like that.

Gonna kill that "first begun" to have it be simple past tense like the other two lines. (I think my insistence on perfect was a matter of unconsciously channelling the final lines of Amazing Grace...)

"No doubt there" was fine and I'm not feeling whatever had gotten me to change it.