levels: more misc. fixes.
**Map03**
The green corridor had some texture alignment and mismatched flat issues; the switch door was overcomplicated for what it was doing and didn't even look better than just using a simple wall switch texture. The corridor itself had a shadow across the middle that had nothing casting it; it has been removed.

Gave the curved terminal wall light grey flats to not stick out as much.


**Map21**
The cage bars along the walls of the lava border corridor should be impassable.


**Map26**
It's possible if you're careful to grab the super shotgun without triggering the lowering panel, causing the player to be potentially lost without any further indication they should move towards the alcove. Since I want to keep this as a foreshadowing of what is possible with stealing the key from the necromancer, there's now also an armour bonus a little further in.

The rising pit was too easy to jump into before it was finished, ruining the effect. I've packed everything closer together and moved one of the hard spectres into the corner outside the pit. The northern bit of the room is also expanded to let the worms move that way.