dehacked: just say "invisibility".
You're not really infiltrating anything with this powerup (GZDoom weirdness notwithstanding); meanwhile, it's generally assumed in the story that the monsters will have some idea that you're around somewhere and they should be shooting in your general inferred direction. Most of them can probably locate you by sound or even smell or possibly even psychically, but they won't have the precision of being able to see you.

There's no real need to qualify the invisibility in this context, and given the options available under the letter I the need for this specfic word is practically a functional requirement.