levels: misc. E3 minor fixes.
E3M1: Rearranged the two "spinal columns" inside the first entrance. The chamber is larger and the columns are not perfectly octagonal (both to align the textures and to get rid of the fake contrast) and one of them is significantly larger than the other.

E3M6: replacing the (not even working right, you could see the floor moving) flat bleed rising painlord with one that warps into that spot. Changed the textures inside the hatchling fireplace so you're seeing fire everywhere from the inside.

E3M7: Slapped some textures onto the otherwise flat bleeding temporary monster blocker. The difference is barely noticeable even if you're deliberately looking for it.