levels: revisit Map11.
In this I try to address two issues:
1. All the stuff, other than mere difficulty, that makes me always want to immediately idclev/idclip out of this map.
2. The Doom 2-ishness of the blue key branch, as per #732.

One common theme in 1 is to ditch or mitigate a lot of gotchas that are basically beaten by dying and memorizing the sequence on restart (or savescum), in favour of things that remain challenging even when you know it's coming. A lot of mandatory direct backtracking has also been mitigated by opening up optional alternate routes.

The name has been changed since everything in this cluster is named with some diegetic description. I always like to imagine the first three doors as actually being fixed seamless interdimensional portals (with the rest of the "real" facility being the final branch) as an explanation for the seemingly unrelated appearance of each branch.

Note that this adds enough content that it creates a savegame buffer overrun. I am proceeding on the basis that [these are not worth caring about](https://www.doomworld.com/forum/topic/92282-savegame-buffer-overruns/).

Detailed notes on changes in the PR.