levels: make Map11 forcefield bidirectional.
The shellbox on the north side of the gate makes things much easier, and being able to cross from the front instead of using the teleporter minimizes your chances of taking damage from either the worm or anything that might be alerted inside the lobby. The only reason why that forcefield bridge could only be crossed north to south was to avoid making things too easy.

This map already being harder than a few much later maps, such forced artificial difficulty is unnecessary.

I've also fixed a few texture issues, most notably the top and bottom of that new secret corridor monitor and the bars on the second yellow key room gargoyle switch.