levels: address map29 artificial difficulty.
If you pistol start this map, on the teleport into the tower you will be forced to get helplessly shredded by serpentipedes while you're switching to the polaric that's right at the destination. A second one has been added on the platform with the backpack that leads to the teleporter, and the current one has been flagged as multiplayer-only.

Right after this is the blue key, which cannot be retrieved without taking damage from the hurtfloor below. There's no reason to be anywhere near this pit after the blue key so the double telepads doesn't even make sense, except to warp the occasional spider into the starting room which just looks silly. To address all these issues, there is only one pad, right next to the rocket launcher switch, which lines are monster blockers, and the blue key platform is now *higher* than the ledge you reach it from so you can see it moving, retreat back into the tower, and go back to the RL platform and reach the telepad without touching the lava.

The medikit with the blue key has been moved into the main ground area in the northwest.

The jump from the wood tower onto the teleport to get into the stone tower is now surrounded by monster blocking lines.

Got rid of two things that seemed intended only to spite the player:
- the sub-shotgun in the green room, which would cause a weapon-switch just as you face the painlord at the top of the lift, is now just two packs of shells.
- all four doors of the minigunner shack are usable. (They are also equally lit, unlike previously when one of them was slightly darker and it wasn't the one that was usable.)

Added a radsuit and medikit to the back of the green building, so that exploring that area isn't merely unmitigatedly punished.

Added a visible light source to each of the indoor areas.