levels: revisit Map01 changes. (#927)
Removed the flat bleed hack in favour of some explicit bright lights at the bottom of the sign.

The gap in the giant doors is now back up to 32 units wide, but there's a visible chain behind them in some implied past attempt to either force them together or prevent them from opening. Either way, it's a midtex that clearly communicates the player shouldn't try to pass through. The slope is specifically chosen to allow the most view of the more detailed side.

The secret armour closet didn't leave any room for the door to slide through. There is now a track cutting through the inside of that room, and the "inside" edge of the door is peeking out a little.

Also finished the door on the outside. Not at all necessary for playing the map as-is, but just in case anyone in the future wanted to do a variant of this map that actually let you go out there...

The AGM logo over the waterfall in the courtyard was weirdly split up by the central orthogonal section's fake contrast so I've split the middle linedef to make a gentler curve.