levels: more minor fixes. (#949)
E1M3
Changed the front entrance into an outdoor train platform. The train is visible from the window behind the blue door as well.
Added some escapes from some slime pits.

E1M7
Jump-proofed rocket launcher secret, the switch behind the bars in the northeast, and the two teleporters that are opened up by approaching the switch in the southeast. Duck-proofed that one rusty bar blocking the way to the northeast area. (If you're playing a mod that lets you duck less than 12 units it's not proof against that.)

E2M1
Pending final replacement of this map, here's a hacky solution to address #864 involving a door trigger and a hidden multiplayer-only barrel. In singleplayer the path beyond the other side of the gap is now blocked by a wall of rubble. At least this should better signpost that the player is not expected to go that way.

E3M7
Made the final door before the painlord stairs stay-open to let the player backtrack later if they want.

E3M8
Got rid of the sound pipes in favour of merging some sectors, fixing the gargoyle textures in the process. Raised the windows into the "snake pit" slightly to line up with the brick texture.

Map16
Addressed #591. The same trigger that "presses the button" now lowers the wall up ahead, revealing an exit portal a second or so after the bossbrain exit would normally activated.

Map17, Map18, Map31
Addressed #867.

Map30
Addressed my comment in #640 by making the platform right in front of the face a secret, so the player can 100% the map. (Not doing anything for E3M8 because you don't get a stats screen for it anyway.) Also dimmed the arena below 255 so depth can be better visible. (Some map geometry has been slightly altered to avoid orthogonal lines where the fake contrast would look bad.)