levels: minor Phase 1 fixes. (#963)
* levels: minor Phase 1 fixes.

This should address all outstanding issues in #698 and then some.

E1M1
Simplified the corridor behind the west blue key entrance.
Removed the serpentipede inside the secret in easy skill, and unflagged it as ambush so you don't need to necessarily *find* the secret to get 100% kills.

E1M7
Addressed #698 by getting rid of the potential soft-lock crusher trap altogether.

E2M8
Addressed #698 by keeping two tripods at all skill levels but making the polaric available immediately, along with a soulsphere and red armour. In the polaric's place in the hard-to-reach room is a second soulsphere.
I've also made the (medium/hard) polaric room door look like a door to reflect its actual use.

E3M6
Addressed #698 by not just marking the illusory wall as a secret line but also the inside of the room not-on-map. Added a light source to the inside of the room but it wasn't really necessary.

E3M7
Addressed #698 by removing the crusher tags for both sector and line.
Also dded some sound blockers to the entrance to the colon.

E4M6
Addressed the texture alignment issues in #698.
As for the descending sector potential crash, that is (a) *literally impossible* to fix in SLADE; and (b) a far too complex sector touching many things to trust not to break all over again on the slightest future adjustment to map geometry. The intended resulting sector type isn't even a hurtfloor, so there's no gameplay effect with the change. I've added a trim around the edges to make it look half decent when it's descended with no texture change.

E4M8
Addressed #698 map line meta.
Added one more escape lift for a slime pit that was still missing one.
Added a SKAG.

* levels: remove testing e4m8 start.

yeah that was where the SKAG is.