levels: address some E2M4 issues. (#966)
Made the lines in the northeast waterfall windows impassable as it is possible to jump through at least the bottom ones in some sourceports.

Made the staircase leading from the first water area back to the start *much* more visible. (Seriously playing this in vanilla with the sun shining through the window IRL this was even harder to find than the red armour secret...)

Replaced nearly all the METAL step faces with AQPIPE09 because they were *really* hard to see in poor lighting conditions.

Changed those insta-walls into something that could plausibly snap shut like that.

Merged a lot of symmetrical detail sectors.