levels: tweak new map09 warehouse.
Fixed a floaty cell on the forklift.

Broke up some of the monotony in the northernmost aisle. The dirty crate blocking the door is now mysteriously moved back up to the rack rather than disappeared entirely.

Also tweaked the textures in that first grey T-corridor: gave the sides something with a seam that at least kindasorta lines up with the near-door seams, and got rid of the dark lines on the diagonal crate.